requirements.txt
fastapi[standard]
bcrypt
asyncpg
redis
PyJWT
python-dotenv
motor